% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_BestEffort.r
\name{alg.be}
\alias{alg.be}
\title{Best-effort algorithm}
\usage{
alg.be(lpo, max.be = 1, max.bp = 0.5, be = 0.5, interval = 1,
  regime = "first")
}
\arguments{
\item{lpo}{Energy demand series as nummeric vector, unit: kWh}

\item{max.be}{Battery's maximum capacity, unit: kWh}

\item{max.bp}{Battery's maximum charing and discharging rate, unit: kW}

\item{be}{Battery's current state of charge; 0 is empty, 1 is full}

\item{interval}{Demand metering interval, unit: minute}

\item{regime}{If "first" (default) our interpretation of original algorithm [1] is used.
Otherwise, the algorith works as described in [2].}
}
\description{
Modifies the energy demand series using a rechargable battery 
and applying best-effort algorithm to obscure actual energy usage.
}
\details{
In this algorithm, the external load is kept 
unchanged unless the battery is too low or too high. 
In either case, the battery will be fully discharged/charged. 
While the battery is empty/full, the external load is set to 
be equal to the demanding load and only kept unchanged when 
the battery is capable of charging/discharging again.
}
\examples{
set.seed(1)
d <- squares(20)
lpo <- d[, 1]/100
lpm1 <- alg.be(lpo)
lpm2 <- alg.be(lpo, regime = "second")
plot(ts(lpo))
lines(ts(lpm1), col = "red")
lines(ts(lpm2), col = "blue")

lpm1 <- alg.be(lpo, max.be = 0.01, max.bp = 10)
lpm2 <- alg.be(lpo, max.be = 0.01, max.bp = 10, regime = "second")
plot(ts(lpo))
lines(ts(lpm1), col = "red")
lines(ts(lpm2), col = "blue")

lpm <- alg.be(lpo, max.bp = 0)
plot(ts(lpo))
lines(ts(lpm), col = "red")
}
\references{
[1] Kalogridis, Georgios, et al. "Privacy for smart meters: 
Towards undetectable appliance load signatures." 2010 First IEEE 
International Conference on Smart Grid Communications. IEEE, 2010.

[2] Yang, W. et al. 2012. Minimizing private data disclosures in the smart grid. 
Proceedings of the 2012 ACM conference on Computer and 
communications security - CCS '12. (2012), 415.
}
