% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_LS.r
\name{alg.stepping}
\alias{alg.stepping}
\title{Lazy stepping algorithm}
\usage{
alg.stepping(lpo, max.be = 1, max.bp = 0.5, be = 0.5, interval = 1,
  regime = "ls1")
}
\arguments{
\item{lpo}{Energy demand series as nummeric vector, unit: kWh}

\item{max.be}{Battery's maximum capacity, unit: kWh}

\item{be}{Battery's current capacity level; 0 is empty, 1 is full}

\item{interval}{Demand metering interval, unit: minute}

\item{min.be}{Battery's minimum capacity, unit: kWh}

\item{max.bc}{Battery's maximum charing rate, unit: kW}

\item{max.bd}{Battery's maximum discharging rate, unit: kW}

\item{mode}{Used to set battery status when charging/discharging 
rate cannot keep up with power demand. Mode 1 has battery charging 
when it's below half-full. Mode 2 has battery charging randomly.}
}
\value{
A moderated energy series
}
\description{
Moderates a energy demand series using a rechargable battery 
and applying lazy stepping algorithm to obscure actual energy usage.
}
\details{
Lazy stepping algorithms try to maintain the external 
load unless it is pushed to change. There are three cases in 
which changes must occur: (1) When maintaining the load results 
in overcharging the battery, in which case the battery must be 
discharged; (2) When maintaining the load results in low battery, 
in which case the battery must be charged; and (3) when power 
demand is either too low or too high for the previous external load. 
We consider two alternatives, which we call LS1 and LS2. 
LS1 charges the battery if it is below half, and discharges 
it otherwise. LS2 randomly charges or discharges the battery.
}
\examples{
set.seed(1)
d <- squares(20)
lpo <- d[, 1]/100


lpm1 <- lazy.stepping.moderate(lpo)
lpm2 <- alg.stepping(lpo)
lpm3 <- lazy.charging.moderate(lpo)
lpm4 <- alg.stepping(lpo, regime = "lc")
set.seed(1)
lpm5 <- random.charging.moderate(lpo)
set.seed(1)
lpm6 <- alg.stepping(lpo, regime = "rc")

plot(ts(lpo))
lines(ts(lpm1), col = "red")
lines(ts(lpm2), col = "blue")
lines(ts(lpm3), col = "red")
lines(ts(lpm4), col = "blue")
lines(ts(lpm5), col = "red")
lines(ts(lpm6), col = "blue")
}
\references{
Yang, Weining, et al. "Minimizing private data 
disclosures in the smart grid." Proceedings of the 2012 
ACM conference on Computer and communications security. ACM, 2012.
}
