% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_BestEffort.r
\name{best.effort.moderate}
\alias{best.effort.moderate}
\title{Best-effort algorithm}
\usage{
best.effort.moderate(lpo, max.be = 1, max.bp = 0.5, be = 0.5,
  interval = 1)
}
\arguments{
\item{lpo}{Energy demand series as nummeric vector, unit: kWh}

\item{max.be}{Battery's maximum capacity, unit: kWh}

\item{max.bp}{Battery's maximum charing and discharging rate, unit: kW}

\item{be}{Battery's current capacity level; 0 is empty, 1 is full}

\item{interval}{Demand metering interval, unit: minute}
}
\description{
Moderates a energy demand series using a rechargable battery 
and applying best-effort algorithm to obscure actual energy usage.
}
\details{
In this algorithm, the external load is kept 
unchanged unless the battery is too low or too high. 
In either case, the battery will be fully discharged/charged. 
While the battery is empty/full, the external load is set to 
be equal to the demanding load and only kept unchanged when 
the battery is capable of charging/discharging again.
}
\examples{
set.seed(1)
d <- squares(20)
lpo <- d[, 1]/100
lpm <- best.effort.moderate(lpo)
plot(ts(lpo))
lines(ts(lpm), col = "red")
}
\references{
Kalogridis, Georgios, et al. "Privacy for smart meters: 
Towards undetectable appliance load signatures." 2010 First IEEE 
International Conference on Smart Grid Communications. IEEE, 2010.
}
