% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_HaarWaveletTransform.r
\name{compress.lpo}
\alias{compress.lpo}
\title{Compress lpo using Haar-wavelet transform.}
\usage{
compress.lpo(lpo, level = 0.99)
}
\arguments{
\item{lpo}{Times series as numeric vector}

\item{level}{Compression level, from 0 to 1}
}
\description{
The time series is transformed using Haar-wavelet transform, 
then the near-zero-values will be rounded to zero to an extend 
and the series will be decoded, creating a new, compressed 
version of the original series.
}
\examples{
set.seed(1)
d <- squares(100)
lpo <- d[, 1]
lpm <- compress.lpo(lpo, 0.9)
plot(ts(lpo))
lines(ts(lpm), col = "red")
}
