% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_mi_hist.R
\name{hist2d}
\alias{hist2d}
\title{2D histogram}
\usage{
hist2d(d)
}
\arguments{
\item{d}{matrix with two columns with quantized variables}
}
\value{
data frame containing joint relative frequency in the first column and marginal 
relative frequencies in the second and in the third
}
\description{
The function omputes 2-dimensional historgamm and two respective one-dimensional ones
}
\examples{
set.seed(1)
d <- squares(100)
hist2d(discr.together(d[, 1], d[, 2], mode = "number"))
}
