% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_mi_hist.R
\name{hist4d}
\alias{hist4d}
\title{4D histogram}
\usage{
hist4d(d)
}
\arguments{
\item{d}{matrix with four columns with quantized variables}
}
\value{
data frame containing joint relative frequency in the first column and 
two-dimensional relative frequencies for
the columns 1:2 and 3:4 of d in the second and the third column correspondingly
}
\description{
computes 4-dimensional historgamm and two two-dimensional histogramms
}
\examples{
set.seed(1)
d <- cbind(squares(10000), squares(10000))
for(i in 1:ncol(d)){
  d[, i] <- discr.sep(d[, i], mode = "number")
}
hist4d(d)
mi.hist.freqs(hist4d(d))
}
