% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_LC.r
\name{lazy.charging.moderate}
\alias{lazy.charging.moderate}
\title{Lazy charging algorithm}
\usage{
lazy.charging.moderate(lpo, max.be = 1, min.be = 0, max.bc = 0.5,
  max.bd = 0.5, be = 0.5, interval = 1)
}
\arguments{
\item{lpo}{Energy demand series as nummeric vector, unit: kWh}

\item{max.be}{Battery's maximum capacity, unit: kWh}

\item{min.be}{Battery's minimum capacity, unit: kWh}

\item{max.bc}{Battery's maximum charing rate, unit: kW}

\item{max.bd}{Battery's maximum discharging rate, unit: kW}

\item{be}{Battery's current capacity level; 0 is empty, 1 is full}

\item{interval}{Demand metering interval, unit: minute}
}
\description{
Moderates a energy demand series using a rechargable battery 
and applying lazy charging algorithm to obscure actual energy usage.
}
\details{
This algorithm tries to maintain charging state 
of battery until the battery is too low or too high. 
One advantage of the LC algorithm is that this reduces 
the number of charge/discharge cycles for the battery. 
One disadvantage of the LC algorithm is that it is generally 
easy to predict charging state of battery at some point.
}
\examples{
set.seed(1)
d <- squares(20)
lpo <- d[, 1]/100
lpm <- lazy.charging.moderate(lpo)
plot(ts(lpo))
lines(ts(lpm), col = "red")
}
\references{
Yang, Weining, et al. "Minimizing private data 
disclosures in the smart grid." Proceedings of the 2012 
ACM conference on Computer and communications security. ACM, 2012.
}
