% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_mi_hist.R
\name{mi.hist.freqs}
\alias{mi.hist.freqs}
\title{Histogram MI estimator for i.i.d. assumption}
\usage{
mi.hist.freqs(freqs)
}
\arguments{
\item{freqs}{table of relative frequencies}
}
\value{
mutual information value
}
\description{
computes mutual information (in nats) provided a table of relative frequencies. 
Assumes that load profiles are i.i.d. samples of random variables
}
\examples{
set.seed(1)
d <- squares(1000)

# if one quantizes jointly
d.together <- discr.together(d[, 1], d[, 2], mode = "number")
h <- hist2d(d.together)
mi.hist.freqs(h)

# if one quantizes separately
d.sep <- d
for(i in 1:ncol(d)){
  d.sep[, i] <- discr.sep(d[, i], mode = "number")
}
h <- hist2d(d.sep)
mi.hist.freqs(h)

# check if any of implementations coincides with the one from the existing package
entropy::mi.empirical(entropy::discretize2d(d[, 1], d[, 2], 20, 20))/log(2)
}
\references{
J. X. Chin, G. Giaconi, T. T. De Rubira, 
D. Gimduz, and G. Hug, "Considering time correlation in the estimation of privacy loss for 
consumers with smart meters," in 20th Power Systems Computation Conference, PSCC 2018.
}
