% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_mi_hist.R
\name{mi.mns.hist}
\alias{mi.mns.hist}
\title{Histogram MI estimator for non-stationary Markov processes}
\usage{
mi.mns.hist(lpo, lpm, features)
}
\arguments{
\item{lpo, lpm}{discretized original and modified load profiles}

\item{features}{the vector of feature values
#' @references J. X. Chin, G. Giaconi, T. T. De Rubira, 
D. Gimduz, and G. Hug, "Considering time correlation in the estimation of privacy loss for 
consumers with smart meters," in 20th Power Systems Computation Conference, PSCC 2018.}
}
\value{
mutual information value
}
\description{
computes mutual information (in nats) provided a table of relative frequencies. 
Assumes that load profiles follow non-stationary first order Markow process with features
}
\examples{
set.seed(1)
d <- squares(250)
lpo <- discr.sep(d[, 1], mode = "number")
lpm <- discr.sep(d[, 2], mode = "number")

features <- c(rep(1, 500), rep(2, 250), rep(3, 250))
mi.mns.hist(lpo, lpm, features)

features <- rep(1:10, each = 100)
mi.mns.hist(lpo, lpm, features)

features <- rep(1:40, each = 25)
mi.mns.hist(lpo, lpm, features)

features <- rep(1:250, each = 4)
mi.mns.hist(lpo, lpm, features)
}
