% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_mi_hist.R
\name{mi.ms.hist}
\alias{mi.ms.hist}
\title{Histogram MI estimator for stationary Markov process assumption}
\usage{
mi.ms.hist(lpo, lpm)
}
\arguments{
\item{lpo, lpm}{discretized original and modified load profiles}
}
\value{
mutual information value
}
\description{
computes mutual information (in nats) provided a table of relative frequencies. 
Assumes that load profiles follow first order stationary Markow process
}
\examples{

set.seed(1)
d <- squares(10000)
lpo <- discr.sep(d[, 1], mode = "number")
lpm <- discr.sep(d[, 2], mode = "number")
mi.ms.hist(lpo, lpm)
}
\references{
J. X. Chin, G. Giaconi, T. T. De Rubira, 
D. Gimduz, and G. Hug, "Considering time correlation in the estimation of privacy loss for 
consumers with smart meters," in 20th Power Systems Computation Conference, PSCC 2018.
}
