% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_mi_knn.R
\name{mi.ms.knn}
\alias{mi.ms.knn}
\title{kNN MI estimator for stationary Markov processes}
\usage{
mi.ms.knn(lpo, lpm, k)
}
\arguments{
\item{lpo, lpm}{original and modified load profiles}

\item{k}{number of nearest neighbor to consider}
}
\value{
mutual information value
}
\description{
computes mutual information between two load profiles (i.e., time series)
under the assumtion of the first-order Markov process
}
\examples{
set.seed(1)
d <- squares(1000)
lpo <- d[, 1]
lpm <- d[, 2]
mi.ms.knn(lpo, lpm, 2)

lpo <- c(1, 1.5, 1.9, 2.1, 2, 1.8, 1.5, 1.5, 1.7, 1.9, 2, 1.9, 1.7, 1.5, 1.2, 1)
lpm <- c(0.7, 2, 1.5, 2.5, 1.6, 1.7, 1.4, 1, 1.9, 1.4, 2.1, 2.7, 1.8, 2.2, 0.6, 1.5)
mi.ms.knn(lpo, lpm, k = 1)
mi.ms.knn(lpo, lpm, k = 4)
}
