% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_NILL.r
\name{nill.moderate}
\alias{nill.moderate}
\title{NILL algorithm}
\usage{
nill.moderate(lpo, max.be = 1, min.be = 0, max.bc = 0.5,
  max.bd = 0.5, be = 0.5, alpha = 0.5, voltage = 110,
  interval = 1)
}
\arguments{
\item{lpo}{Energy demand series as nummeric vector, unit: kWh}

\item{max.be}{Battery's maximum capacity, unit: kWh}

\item{min.be}{Battery's minimum capacity, unit: kWh}

\item{max.bc}{Battery's maximum charing rate, unit: kW}

\item{max.bd}{Battery's maximum discharging rate, unit: kW}

\item{be}{Battery's current capacity level; 0 is empty, 1 is full}

\item{alpha}{Coefficient used to reset external load constant}

\item{voltage}{Used to convert between Amp and Watt, unit: vol}

\item{interval}{Demand metering interval, unit: minute}
}
\description{
Moderates a energy demand series using a rechargable battery 
and applying NILL algorithm to obscure actual energy usage.
}
\details{
The NILL algorithm has three states. In stable state (ST), 
NILL tries to maintain external load at some value Kst. When the 
battery is full, it moves to the high recovery state (HR), in 
which the external load will be set lower than the demand and 
the battery will be gradually discharged. The opposite case takes 
place when in stable state the battery is too low.
}
\examples{
set.seed(1)
d <- squares(20)
lpo <- d[, 1]/100
lpm <- nill.moderate(lpo)
plot(ts(lpo))
lines(ts(lpm), col = "red")
}
\references{
McLaughlin, Stephen, Patrick McDaniel, and William Aiello. 
"Protecting consumer privacy from electric load monitoring." 
Proceedings of the 18th ACM conference on Computer and 
communications security. ACM, 2011

Yang, Weining, et al. "Minimizing private data 
disclosures in the smart grid." Proceedings of the 2012 ACM 
conference on Computer and communications security. ACM, 2012.
}
