% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_cs.R
\name{priv.cs}
\alias{priv.cs}
\title{Cluster similarity}
\usage{
priv.cs(lpo, lpm, n = -1)
}
\arguments{
\item{lpo, lpm}{original and modified load profiles}

\item{n}{number of clusters used. Should be greater than 2. If not given an optimized number will be computed.}
}
\value{
privacy measure: cluster similarity. If lpm is constant, the similarity is zero.
}
\description{
Calculates privacy measure based on cluster similarity.
}
\examples{
set.seed(1)
d <- squares(1000)
lpo <- d[, 1]
lpm <- d[, 2]
priv.cs(lpo, lpm)
priv.cs(lpo, lpm, n = 20)

lpo <- c(1, 1.5, 1.9, 2.1, 2, 1.8, 1.5, 1.5, 1.7, 1.9, 2, 1.9, 1.7, 1.5, 1.2, 1)
lpm <- c(0.7, 2, 1.5, 2.5, 1.6, 1.7, 1.4, 1, 1.9, 1.4, 2.1, 2.7, 1.8, 2.2, 0.6, 1.5)
priv.cs(lpo, lpm)
}
\references{
Kalogridis, Georgios, et al. "Elecprivacy: Evaluating the privacy protection of electricity 
management algorithms." IEEE Transactions on Smart Grid 2.4 (2011): 750-758.
}
