% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_dc.R
\name{priv.dc}
\alias{priv.dc}
\title{Privacy measure: "Coefficient of determination"}
\usage{
priv.dc(lpo, lpm, n = 10, regime = c("reg", "lpo"))
}
\arguments{
\item{lpo, lpm}{original and modified load profiles}

\item{n}{number of tinestamps to consider in the search of best alingment}

\item{regime}{character. If "reg" the linear regression is fitted to derive the predicted values as in [1].
if "lpo", the predicted values coincide with the original load profile ("user load") as in [2].}
}
\value{
privacy measure: determination coefficient
}
\description{
Calculates privacy measure called "Coeficient of determination". 
This coincides with coefficient of determination betweed shifted
vectors in its common meaning only if option "reg" is used.
}
\examples{
set.seed(1)
d <- squares(1000)
lpo <- d[, 1]
lpm <- d[, 2]
priv.dc(lpo, lpm, n = 20, regime = "reg")
priv.dc(lpo, lpm, n = 20, regime = "lpo")
}
\references{
[1] G. Kalogridis, S. Z. Denic, T. Lewis, and R. Cepeda,
"Privacy protection system and metrics for hiding electrical events,"
International Journal of Security and Networks, vol. 6, no. 1, p. 14, 2011.

[2] Chen, Zhi, and Lei Wu. "Residential appliance DR energy management 
with electric privacy protection by online stochastic optimization." 
IEEE Transactions on Smart Grid 4.4 (2013): 1861-1869
}
