% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_er_knn.R
\name{priv.er.knn}
\alias{priv.er.knn}
\title{Nearest neighbor estimator entropy ratio}
\usage{
priv.er.knn(lpo, lpm, k = 2, regime = c("zeros", "nozeros"),
  noise = 10^-8)
}
\arguments{
\item{lpo, lpm}{original and modified load profiles}

\item{k}{number of nearest neigbors}

\item{regime}{character. If "zeros", zero values are included in calculation; if "nozeros" - not included.}

\item{noise}{the amplitude of the noise to add in case of duplicated points}
}
\value{
knn entropy ratio estimate
}
\description{
Calculates privacy as the ratio of load provile entropy after 
and before modification.
}
\examples{
set.seed(1)
d <- squares(1000)
lpo <- d[, 1]
lpm <- d[, 2]
priv.er.knn(lpo, lpm, k = 1, regime = "zeros")
priv.er.knn(lpo, lpm, k = 2, regime = "zeros")
priv.er.knn(lpo, lpm, k = 2, regime = "nozeros")

lpo <- c(1, 1.5, 1.9, 2.1, 2, 1.8, 1.5, 1.5, 1.7, 1.9, 2, 1.9, 1.7, 1.5, 1.2, 1)
lpm <- c(0.7, 2, 1.5, 2.5, 1.6, 1.7, 1.4, 1, 1.9, 1.4, 2.1, 2.7, 1.8, 2.2, 0.6, 1.5)
priv.er.knn(lpo, lpm, k = 1)
priv.er.knn(lpo, lpm, k = 4)
}
\references{
S. McLaughlin, P. McDaniel, and W. Aiello, 
"Protecting consumer privacy from electric load monitoring," 
Proceedings of the 18th ACM conference on Computer and 
communications security - CCS '11, p. 87, 2011.
}
