% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_fm.R
\name{priv.fm}
\alias{priv.fm}
\title{Relative feature mass}
\usage{
priv.fm(lpo, lpm, thr = 0, regime = c("fm", "rfm", "ed"))
}
\arguments{
\item{lpo, lpm}{original and modified load profiles}

\item{thr}{threshold value. The value of load profile exceeding thr is a feature}

\item{regime}{character specifying output. "fm" - feature mass, "rfm - relative
feature mass, "ed" - event detection accuracy.}
}
\value{
value of feature mass based privacy measure
}
\description{
Calculates privacy measure based on feature mass.
}
\examples{
set.seed(1)
d <- squares(1000)
lpo <- d[, 1]
lpm <- d[, 2]
priv.fm(lpo, lpm, thr = 1.5, regime = "ed")
priv.fm(lpo, lpm, thr = 1.5, regime = "fm")
priv.fm(lpo, lpm, thr = 1.5, regime = "rfm")

lpo <- c(1, 1.5, 1.9, 2.1, 2, 1.8, 1.5, 1.5, 1.7, 1.9, 2, 1.9, 1.7, 1.5, 1.2, 1)
lpm <- c(0.7, 2, 1.5, 2.5, 1.6, 1.7, 1.4, 1, 1.9, 1.4, 2.1, 2.7, 1.8, 2.2, 0.6, 1.5)
priv.fm(lpo, lpm, thr = 1.6, regime = "ed")
priv.fm(lpo, lpm, thr = 1.6, regime = "fm")
priv.fm(lpo, lpm, thr = 1.6, regime = "rfm")
}
\references{
Relative feature mass: S. McLaughlin, P. McDaniel, and W. Aiello,
"Protecting consumer privacy from electric load monitoring," 
Proceedings of the 18th ACM conference on Computer and 
communications security - CCS '11, p. 87, 2011.

Feature mass: Yang, Lei, et al. "Cost-effective and privacy-preserving 
energy management for smart meters." 
IEEE Transactions on Smart Grid 6.1 (2014): 486-495.

Event detection: Zhao, Jing, et al. "Achieving differential privacy 
of data disclosure in the smart grid." IEEE INFOCOM 2014-IEEE 
Conference on Computer Communications. IEEE, 2014.
}
