% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_kd.R
\name{priv.kd}
\alias{priv.kd}
\title{K-Divergence Privacy Measure}
\usage{
priv.kd(lpo, lpm, num.bins = 20, bin.size = 0.01, mode = "number")
}
\arguments{
\item{lpo, lpm}{original and modified load profiles}

\item{num.bins}{number of bins}

\item{bin.size}{size of bins}

\item{mode}{if "number", load profile is split into num.bins of equidistant bins;
if "size", bin.size is used for defining the size of each bin;
if "mulaw", mu-law discretization scheme is used.}
}
\value{
K-Divergence value
}
\description{
K-Divergence Privacy Measure
}
\examples{
set.seed(1)
d <- squares(1000)
lpo <- d[, 1]
lpm <- d[, 2]
priv.kd(lpo, lpm)
}
