% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_kl_hist.R
\name{priv.kl.hist}
\alias{priv.kl.hist}
\title{Histogram estimator for KL divergence}
\usage{
priv.kl.hist(lpo, lpm, num.bins = 20, bin.size = 0.01,
  mode = "number")
}
\arguments{
\item{lpo, lpm}{original and modified load profiles}

\item{num.bins}{number of bins}

\item{bin.size}{size of bins}

\item{mode}{if "number", load profile is split into num.bins of equidistant bins;
if "size", bin.size is used for defining the size of each bin.}
}
\value{
KL divergence estimate
}
\description{
The function estimates Kullback-Leibler divergence bethween two distributions estimated
from time series data (load profiles). Preliminary quantizes their values.
}
\examples{
set.seed(1)
d <- squares(1000)
lpo <- d[, 1]
lpm <- d[, 2]
priv.kl.hist(lpo, lpm)

lpo <- c(1, 1.5, 1.9, 2.1, 2, 1.8, 1.5, 1.5, 1.7, 1.9, 2, 1.9, 1.7, 1.5, 1.2, 1)
lpm <- c(0.7, 2, 1.5, 2.5, 1.6, 1.7, 1.4, 1, 1.9, 1.4, 2.1, 2.7, 1.8, 2.2, 0.6, 1.5)
priv.kl.hist(lpo, lpm, num.bins = 2, mode = "number")
}
\references{
Kalogridis, Georgios, et al. 
"Privacy for smart meters: Towards undetectable appliance load signatures." 
2010 First IEEE International Conference on Smart Grid Communications. 
IEEE, 2010.
}
