% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_mi_hist.R
\name{priv.mi.hist}
\alias{priv.mi.hist}
\title{Histogram MI estimator for different assumptions}
\usage{
priv.mi.hist(lpo, lpm, regime = c("iid", "ms", "mns", "bin"),
  num.bins = 20, bin.size = 0.01, mode = "number", features = NA)
}
\arguments{
\item{lpo, lpm}{original and modified load profiles}

\item{regime}{assumption on the type of process; "iid" - 
independently identically distributed, "ms" - first order Markov stationary,
"mns" - first order Markov non-stationary with features,
"bin" - the MI with i.i.d. assumption and two bins (num.bins = 2)}

\item{num.bins}{number of bins}

\item{bin.size}{size of bins}

\item{mode}{if "number", load profile is split into num.bins of equidistant bins;
if "size", bin.size is used for defining the size of each bin.}

\item{features}{features for non-stationary Markow process}
}
\value{
mutual information value
}
\description{
computes mutual information (in nats) between two load profiles (i.e., time series)
under various assumptions and variables transformations (= choices of binning schemes)
}
\examples{
set.seed(1)
d <- squares(1000)
lpo <- d[, 1]
lpm <- d[, 2]
priv.mi.hist(lpo, lpm, regime = "iid", num.bins = 20)
# binary version:
priv.mi.hist(lpo, lpm, regime = "iid", num.bins = 2)
priv.mi.hist(lpo, lpm, regime = "bin")

priv.mi.hist(lpo, lpm, regime = "ms", num.bins = 20)

set.seed(1)
d <- squares(250)
lpo <- d[, 1]
lpm <- d[, 2] 
features <- c(rep(1, 500), rep(2, 250), rep(3, 250))
priv.mi.hist(lpo, lpm, regime = "mns", features = features)
}
\references{
J. X. Chin, G. Giaconi, T. T. De Rubira, 
D. Gimduz, and G. Hug, "Considering time correlation in the estimation of privacy loss for 
consumers with smart meters," in 20th Power Systems Computation Conference, PSCC 2018.
}
