% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_mi_knn.R
\name{priv.mi.knn}
\alias{priv.mi.knn}
\title{kNN MI estimator for different assumptions}
\usage{
priv.mi.knn(lpo, lpm, regime = c("iid", "ms", "mns"), k = 2,
  features = rep(1, length(lpo)), noise = 10^-8)
}
\arguments{
\item{lpo, lpm}{original and modified load profiles}

\item{regime}{assumption on the type of process; "iid" - 
independently identically distributed, "ms" - first order Markov stationary,
"mns" - first order Markov non-stationary with features}

\item{k}{number of nearest neighbor to consider}

\item{features}{features for non-stationary Markow process}

\item{noise}{the amplitude of the noise to add in case of duplicated points}
}
\value{
mutual information value
}
\description{
computes mutual information between two load profiles (i.e., time series)
under various assumptions and variables transformations
}
\examples{
set.seed(1)
d <- squares(100)
features <- rep(1:10, each = 40)
priv.mi.knn(d[, 1], d[, 2], regime = "iid", k = 2, features = features)
priv.mi.knn(d[, 1], d[, 2], regime = "ms", k = 2, features = features)
priv.mi.knn(d[, 1], d[, 2], regime = "mns", k = 2, features = features) 

features <- rep(1:4, each = 4)
lpo <- c(1, 1.5, 1.9, 2.1, 2, 1.8, 1.5, 1.5, 1.7, 1.9, 2, 1.9, 1.7, 1.5, 1.2, 1)
lpm <- c(0.7, 2, 1.5, 2.5, 1.6, 1.7, 1.4, 1, 1.9, 1.4, 2.1, 2.7, 1.8, 2.2, 0.6, 1.5)
priv.mi.knn(lpo, lpm, regime = "iid", k = 2, features = features)
priv.mi.knn(lpo, lpm, regime = "ms", k = 2, features = features)
priv.mi.knn(lpo, lpm, regime = "mns", k = 2, features = features) 
}
\references{
J. X. Chin, G. Giaconi, T. T. De Rubira, D. Gimduz, and G. Hug, 
"Considering time correlation in the estimation of privacy loss for 
consumers with smart meters," in 20th Power Systems Computation 
Conference, PSCC 2018, 2018.
}
