% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_rc.R
\name{priv.rc}
\alias{priv.rc}
\title{Reconstruction-based privacy measure}
\usage{
priv.rc(lpo, lpm, regime = c("v", "w"))
}
\arguments{
\item{lpo, lpm}{original and modified load profiles}

\item{regime}{if "v" value leaks is used; if "w" - wavelet filtering is used}
}
\value{
the value of the reconstruction privacy measure
}
\description{
Reconstruction-based privacy measure
}
\examples{
lpo <- c(1, 1.5, 1.9, 2.1, 2, 1.8, 1.5, 1.5, 1.7, 1.9, 2, 1.9, 1.7, 1.5, 1.2, 1)
lpm <- c(0.7, 2, 1.5, 2.5, 1.6, 1.7, 1.4, 1, 1.9, 1.4, 2.1, 2.7, 1.8, 2.2, 0.6, 1.5)
priv.rc(lpo, lpm, regime = "v")
priv.rc(lpm, lpo, regime = "v")
priv.rc(lpo, lpm, regime = "w")
priv.rc(lpm, lpo, regime = "w")
}
