% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_rc.R
\name{priv.rc.v}
\alias{priv.rc.v}
\title{Find correlation between the perturbed time series and original one.}
\usage{
priv.rc.v(lpo, lpm)
}
\arguments{
\item{lpo}{Original time series}

\item{lpm}{Modified time series}
}
\value{
Time series representing correlation.
}
\description{
Find correlation between the perturbed time series and original one.
}
\examples{
lpo <- c(1, 1.5, 1.9, 2.1, 2, 1.8, 1.5, 1.5, 1.7, 1.9, 2, 1.9, 1.7, 1.5, 1.2, 1)
lpm <- c(0.7, 2, 1.5, 2.5, 1.6, 1.7, 1.4, 1, 1.9, 1.4, 2.1, 2.7, 1.8, 2.2, 0.6, 1.5)
priv.rc.v(lpo, lpm)
priv.rc.v(lpm, lpo)

set.seed(1)
d <- squares(17)
priv.rc.v(d[, 1], d[, 2])
priv.rc.v(d[, 2], d[, 1])
}
