% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_rc.R
\name{priv.rc.w}
\alias{priv.rc.w}
\title{Attempt to reconstruct the original profile using wavelet filtering}
\usage{
priv.rc.w(lpo, lpm)
}
\arguments{
\item{lpo, lpm}{original and modified load profiles}
}
\value{
Predicted original time series
}
\description{
Attempt to reconstruct the original profile using wavelet filtering
}
\examples{
lpo <- c(1, 1.5, 1.9, 2.1, 2, 1.8, 1.5, 1.5, 1.7, 1.9, 2, 1.9, 1.7, 1.5, 1.2, 1)
lpm <- c(0.7, 2, 1.5, 2.5, 1.6, 1.7, 1.4, 1, 1.9, 1.4, 2.1, 2.7, 1.8, 2.2, 0.6, 1.5)
priv.rc.w(lpo, lpm)
priv.rc.w(lpm, lpo)

set.seed(1)
d <- squares(17)
priv.rc.w(d[, 1], d[, 2])
priv.rc.w(d[, 2], d[, 1])
}
