% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priv_tvd.R
\name{priv.tvd}
\alias{priv.tvd}
\title{Histogramm estimator for total variation distance}
\usage{
priv.tvd(lpo, lpm, num.bins = 20, bin.size = 0.01, mode = "number")
}
\arguments{
\item{lpo, lpm}{original and modified load profiles}

\item{num.bins}{number of bins}

\item{bin.size}{size of bins}

\item{mode}{if "number", load profile is split into num.bins of equidistant bins;
if "size", bin.size is used for defining the size of each bin.}
}
\value{
total variation distance
}
\description{
The function stimates total variation diatance bethween two distributions estimated
from time series data (load profiles). Preliminary quantizes their values.
}
\examples{
set.seed(1)
d <- squares(10000)
priv.tvd(d[, 1], d[, 2])
}
\references{
G. Kalogridis, R. Cepeda, S. Z. Denic, T. Lewis, and C. Efthymiou, 
"ElecPrivacy: Evaluating the privacy protection of electricity 
management algorithms," IEEE Transactions on Smart Grid, 
vol. 2, no. 4, pp. 750-758, 2011.
}
