% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_RC.r
\name{random.charging.moderate}
\alias{random.charging.moderate}
\title{Random charging algorithm}
\usage{
random.charging.moderate(lpo, max.be = 1, min.be = 0, max.bc = 0.5,
  max.bd = 0.5, be = 0.5, interval = 1)
}
\arguments{
\item{lpo}{Energy demand series as nummeric vector, unit: kWh}

\item{max.be}{Battery's maximum capacity, unit: kWh}

\item{min.be}{Battery's minimum capacity, unit: kWh}

\item{max.bc}{Battery's maximum charing rate, unit: kW}

\item{max.bd}{Battery's maximum discharging rate, unit: kW}

\item{be}{Battery's current capacity level; 0 is empty, 1 is full}

\item{interval}{Demand metering interval, unit: minute}
}
\description{
Moderates a energy demand series using a rechargable battery
and applying random charging algorithm to obscure actual energy usage.
}
\details{
In this algorithm, the charging state of the
battery is set randomly each time a load demand comes in.
}
\examples{
set.seed(1)
d <- squares(20)
lpo <- d[, 1]/100
lpm <- random.charging.moderate(lpo)
plot(ts(lpo))
lines(ts(lpm), col = "red")
}
\references{
Yang, Weining, et al. "Minimizing private data
disclosures in the smart grid." Proceedings of the 2012 ACM
conference on Computer and communications security. ACM, 2012.
}
