% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oth_rmix.R
\name{rmix}
\alias{rmix}
\title{Mixes outcones for non-stationary Markov process with features}
\usage{
rmix(d1, d2)
}
\arguments{
\item{d1, d2}{two column tables for outcomes of two consequitive periods of load profiles defined by 
feature values}
}
\value{
the table of outcomes
}
\description{
For two consecutive feature values produces the combinations of different outcomes for estimating 
4-dimensional probabilities
}
\examples{
x <- c("a", "b", "c")
y <- 1:3
d1 <- cbind(x, y)
x <- c("d", "e", "f")
y <- 4:6
d2 <- cbind(x, y)
rmix(d1, d2)
}
\references{
J. X. Chin, G. Giaconi, T. T. De Rubira, D. Gimduz, and G. Hug, 
"Considering time correlation in the estimation of privacy loss for consumers with smart meters," 
in 20th Power Systems Computation Conference, PSCC 2018, 2018.
}
