% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_squares.R
\name{squares}
\alias{squares}
\title{Synthetic data with known MI}
\usage{
squares(N, sizes = c(2, 1, 1), p = 1)
}
\arguments{
\item{N}{number of points in the smallest square. The total number op points 
generated will be N*sum(sizes)*p}

\item{sizes}{a vector with side sizes of each square. The number of points generated
from each square is proportional to its side's size}

\item{p}{number of repetitions}
}
\value{
2D time series of length N*p*sum(sizes)
}
\description{
Generates the data from a set of squares with known theoretical value of Mutual Information
}
\examples{
set.seed(1)
d <- squares(30, c(2, 1), 3)
plot(d)
plot(ts(d[, 1]))
plot(d)
}
